refactor(package): change the package name from con-fi to contra
Con-fi was borrowed from chroma-fi and it seems like it's not coherent with what this package is
doing.
Contra seem natural as it joins console 'con' (this is a console function) and trace 'tra'
(this package sort of gives a short trace of the log source).

BREAKING CHANGE: The name is now deprecated in any dependants, though there is none at the moment, I
have not published this to npm yet.